# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Endorphins API",
    "version": "0.3.0"
  },
  "paths": {
    "/api/v1/workouts": {
      "post": {
        "operationId": "generateWorkout",
        "summary": "Generate a full-body workout from the original exercise catalogue",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "JSON required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Unsupported duration or level",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Generation unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "201": {
            "description": "Generated workout saved to the authenticated account. Times are estimates.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workout"
                }
              }
            },
            "headers": {
              "Location": {
                "description": "Saved workout resource URL",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Key already used for different generation input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Optional owner-scoped request key. Same key and duration/level return the original saved workout (201); a different input returns 409. Retained for the life of the workout.",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "pattern": "^[!-~]+$"
            }
          }
        ]
      },
      "get": {
        "operationId": "listWorkouts",
        "summary": "Search your complete saved library with stable cursor pagination",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 512
            },
            "description": "Opaque cursor scoped to owner and normalized filters/order; changing filters requires a new first page."
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 20
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "Literal case-insensitive search across all owned workout focus, block, and exercise names. Trimmed; at most 200 Unicode characters.",
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          },
          {
            "name": "level",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 5
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "newest",
                "shortest"
              ],
              "default": "newest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Your saved workouts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workouts/{id}": {
      "get": {
        "operationId": "getWorkout",
        "summary": "Read a workout owned by your account",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Saved workout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workout"
                }
              }
            }
          },
          "404": {
            "description": "Workout not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me": {
      "get": {
        "operationId": "currentUser",
        "summary": "Resolve your Clerk identity to an Endorphins user",
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateAccount",
        "summary": "Set the default level and finish onboarding",
        "responses": {
          "200": {
            "description": "Set the default level and finish onboarding",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Retry key conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid selection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/workouts/summary": {
      "get": {
        "operationId": "summarizeWorkouts",
        "summary": "Summarize all owned plans matching the active q/level filters, independent of pagination",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Literal case-insensitive search across all owned workout focus, block, and exercise names. Trimmed; at most 200 Unicode characters.",
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          },
          {
            "name": "level",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plan counts and estimated minutes across the matching library. These are generated plans, not completed training.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutSummary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid filters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/export": {
      "get": {
        "operationId": "exportAccount",
        "summary": "Download a consistent snapshot of your account and all saved workouts",
        "responses": {
          "200": {
            "description": "Private account data download",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountExport"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Account no longer exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Export unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/webhooks/clerk": {
      "post": {
        "operationId": "clerkWebhook",
        "security": [],
        "summary": "Verified Svix user.deleted delivery; erase data and retain a subject digest against stale sessions",
        "parameters": [
          {
            "name": "svix-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "svix-timestamp",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "svix-signature",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "type",
                  "data"
                ],
                "properties": {
                  "type": {
                    "type": "string"
                  },
                  "data": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Verified event processed, including harmless duplicate deletions or ignored event types"
          },
          "400": {
            "description": "Invalid signature, timestamp, or payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Erasure failed; delivery must be retried",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Webhook is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workouts/{id}/completions": {
      "post": {
        "operationId": "completeWorkout",
        "summary": "Confirm that you did a saved workout",
        "responses": {
          "201": {
            "description": "Confirm that you did a saved workout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Completion"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Retry key conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid selection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ]
      }
    },
    "/api/v1/completions/{id}": {
      "delete": {
        "operationId": "undoCompletion",
        "summary": "Undo one completion without changing the saved plan",
        "responses": {
          "200": {
            "description": "Undo one completion without changing the saved plan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UndoCompletion"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Retry key conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid selection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/activity": {
      "get": {
        "operationId": "getActivity",
        "summary": "Read completed activity in your local time zone",
        "responses": {
          "200": {
            "description": "Read completed activity in your local time zone",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Activity"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Retry key conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid selection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "timezone",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 100
            },
            "example": "America/New_York"
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "GenerateRequest": {
        "type": "object",
        "properties": {
          "durationMinutes": {
            "type": "integer",
            "minimum": 30,
            "maximum": 120
          },
          "level": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          }
        },
        "required": [
          "durationMinutes",
          "level"
        ],
        "additionalProperties": false
      },
...
      "Block": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "enum": [
              "legs",
              "upper body",
              "core"
            ]
          },
          "sets": {
            "type": "integer",
            "minimum": 1
          },
          "estimatedMinutes": {
            "type": "integer",
            "minimum": 1
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "easy",
              "medium",
              "hard"
            ]
          },
          "exercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Exercise"
            },
            "minItems": 1
          }
        },
        "required": [
          "name",
          "sets",
          "estimatedMinutes",
          "difficulty",
          "exercises"
        ],
        "additionalProperties": false
      },
      "Workout": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "level": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "requestedMinutes": {
            "type": "integer",
            "minimum": 30,
            "maximum": 120
          },
          "estimatedMinutes": {
            "type": "integer",
            "minimum": 1,
            "maximum": 120
          },
          "warmupMinutes": {
            "type": "integer",
            "minimum": 0,
            "maximum": 5
          },
          "focus": {
            "type": "string",
            "enum": [
              "legs",
              "upper body",
              "core"
            ]
          },
          "blocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Block"
            },
            "minItems": 3,
            "maxItems": 3
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "level",
          "requestedMinutes",
          "estimatedMinutes",
          "warmupMinutes",
          "focus",
          "blocks",
          "createdAt"
        ],
        "additionalProperties": false
      },
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "message",
          "code",
          "requestId"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        }
      },
      "User": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "clerkUserId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "defaultLevel": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "onboardingCompletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "clerkUserId",
          "createdAt",
          "defaultLevel",
          "onboardingCompletedAt"
        ]
      },
      "WorkoutPage": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Workout"
            }
          },
          "nextCursor": {
            "type": "string"
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
      "WorkoutSummary": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "count",
          "plannedMinutes",
          "averageMinutes",
          "levels"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "minimum": 0
          },
          "plannedMinutes": {
            "type": "integer",
            "minimum": 0
          },
          "averageMinutes": {
            "type": "number",
            "minimum": 0
          },
          "levels": {
            "type": "array",
            "minItems": 5,
            "maxItems": 5,
            "items": {
              "type": "object",
              "required": [
                "level",
                "count"
              ],
              "additionalProperties": false,
              "properties": {
                "level": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 5
                },
                "count": {
                  "type": "integer",
                  "minimum": 0
                }
              }
            }
          }
        }
      },
      "AccountExport": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "version",
          "exportedAt",
          "user",
          "workouts",
          "completions"
        ],
        "properties": {
          "version": {
            "type": "integer",
            "enum": [
              2
            ]
          },
          "exportedAt": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "workouts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Workout"
            }
          },
          "completions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Completion"
            }
          }
        }
      },
      "UpdateAccountRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "defaultLevel": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "completeOnboarding": {
            "type": "boolean"
          }
        },
        "required": [
          "defaultLevel"
        ]
      },
      "Completion": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "workoutId": {
            "type": "string"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time"
          },
          "undoneAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "level": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "focus": {
            "type": "string",
            "enum": [
              "legs",
              "upper body",
              "core"
            ]
          }
        },
        "required": [
          "id",
          "workoutId",
          "completedAt",
          "undoneAt",
          "level",
          "focus"
        ]
      },
...
      "Activity": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "completedCount": {
            "type": "integer",
            "minimum": 0
          },
          "activeDaysThisWeek": {
            "type": "integer",
            "minimum": 0
          },
          "weeks": {
            "type": "array",
            "minItems": 4,
            "maxItems": 4,
            "items": {
              "$ref": "#/components/schemas/ActivityWeek"
            }
          },
          "recent": {
            "type": "array",
            "maxItems": 5,
            "items": {
              "$ref": "#/components/schemas/Completion"
            }
          }
        },
        "required": [
          "completedCount",
          "activeDaysThisWeek",
          "weeks",
          "recent"
        ]
      },
      "UndoCompletion": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "undone": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "undone"
        ]
      }
    },
    "securitySchemes": {
      "clerkSession": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "clerkSession": []
    }
  ]
}